Package renaming, time in logs, named constant for exit code
