--- code/infix-to-postfix-notation-converter/View/src/main/java/ru/unn/agile/InfixToPostfixNotationConverter/view/Main.java
@@ -1,21 +1,0 @@
-package ru.unn.agile.InfixToPostfixNotationConverter.view;
-
-import javafx.application.Application;
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Parent;
-import javafx.scene.Scene;
-import javafx.stage.Stage;
-
-public class Main extends Application {
-    @Override
-    public void start(final Stage primaryStage) throws Exception {
-        Parent root = FXMLLoader.load(getClass().getResource("Converter.fxml"));
-        primaryStage.setTitle("Converter");
-        primaryStage.setScene(new Scene(root));
-        primaryStage.show();
-    }
-
-    public static void main(final String[] args) {
-        launch(args);
-    }
-}
