+++ code/infix-to-postfix-notation-converter/View/src/main/resources/ru/unn/agile/infixtopostfixnotationconverter/view/Converter.fxml
@@ -1,0 +1,68 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.ScrollPane?>
+<?import javafx.scene.control.TextField?>
+<?import javafx.scene.layout.AnchorPane?>
+<?import javafx.scene.text.Font?>
+<?import javafx.scene.text.Text?>
+
+<?import javafx.scene.control.ListView?>
+<AnchorPane prefHeight="509.0" prefWidth="807.0" xmlns="http://javafx.com/javafx/8.0.192" xmlns:fx="http://javafx.com/fxml/1" fx:controller="ru.unn.agile.infixtopostfixnotationconverter.view.Converter">
+   <children>
+      <TextField fx:id="infixExpression" layoutX="14.0" layoutY="105.0" prefHeight="25.0" prefWidth="366.0">
+         <font>
+            <Font size="20.0" />
+         </font>
+      </TextField>
+      <Text layoutX="170.0" layoutY="27.740234375" strokeType="OUTSIDE" strokeWidth="0.0" text="Infix To Postfix Notation Converter" textAlignment="CENTER" wrappingWidth="807.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="4.0">
+         <font>
+            <Font size="22.0" />
+         </font>
+      </Text>
+      <TextField fx:id="postfixExpression" editable="false" layoutX="427.0" layoutY="105.0" prefHeight="25.0" prefWidth="366.0">
+         <font>
+            <Font size="20.0" />
+         </font>
+      </TextField>
+      <TextField fx:id="expressionResult" editable="false" layoutX="427.0" layoutY="211.0" prefHeight="25.0" prefWidth="366.0">
+         <font>
+            <Font size="20.0" />
+         </font>
+      </TextField>
+      <Text fx:id="status" fill="#00000074" layoutX="14.0" layoutY="252.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Status:" wrappingWidth="366.0">
+         <font>
+            <Font size="20.0" />
+         </font>
+      </Text>
+      <Button fx:id="convertAndCalculateButton" layoutX="14.0" layoutY="167.0" mnemonicParsing="false" prefHeight="44.0" prefWidth="366.0" text="Convert &amp; Calculate">
+         <font>
+            <Font size="20.0" />
+         </font>
+      </Button>
+      <Text layoutX="14.0" layoutY="93.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Input infix expression here:">
+         <font>
+            <Font size="20.0" />
+         </font>
+      </Text>
+      <Text layoutX="584.0" layoutY="92.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Postfix notation output:" textAlignment="RIGHT">
+         <font>
+            <Font size="20.0" />
+         </font>
+      </Text>
+      <Text layoutX="610.0" layoutY="197.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Result of expression:" textAlignment="RIGHT">
+         <font>
+            <Font size="20.0" />
+         </font>
+      </Text>
+      <ScrollPane layoutX="14.0" layoutY="286.0">
+        <content>
+          <AnchorPane minHeight="0.0" minWidth="0.0" prefHeight="200.0" prefWidth="775.0">
+               <children>
+                  <ListView fx:id="log" layoutX="28.0" layoutY="14.0" prefHeight="200.0" prefWidth="775.0" AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0" />
+               </children>
+            </AnchorPane>
+        </content>
+      </ScrollPane>
+   </children>
+</AnchorPane>
