--- code/infix-to-postfix-notation-converter/View/src/test/java/ru/unn/agile/InfixToPostfixNotationConverter/view/FakeViewTests.java
@@ -1,13 +1,0 @@
-package ru.unn.agile.InfixToPostfixNotationConverter.view;
-
-import org.junit.Ignore;
-import org.junit.Test;
-
-import static org.junit.Assert.fail;
-
-public class FakeViewTests {
-    @Ignore @Test
-    public void fakeTest() {
-        fail("No tests for view");
-    }
-}
