--- code/infix-to-postfix-notation-converter/ViewModel/src/main/java/ru/unn/agile/InfixToPostfixNotationConverter/viewmodel/ILogger.java
@@ -1,8 +1,0 @@
-package ru.unn.agile.InfixToPostfixNotationConverter.viewmodel;
-
-import java.util.List;
-
-public interface ILogger extends AutoCloseable {
-    void log(String message);
-    List<String> getMessages();
-}
