+++ code/infix-to-postfix-notation-converter/ViewModel/src/main/java/ru/unn/agile/infixtopostfixnotationconverter/viewmodel/ILogger.java
@@ -1,0 +1,8 @@
+package ru.unn.agile.infixtopostfixnotationconverter.viewmodel;
+
+import java.util.List;
+
+public interface ILogger extends AutoCloseable {
+    void log(String message);
+    List<String> getMessages();
+}
